Add retries to HTTP client
This uses the urllib3 Retry to add retries with back-off to requests
to the API-server that error. It will retry errors with the listed
statuses, for all HTTP methods. If the response contains a Retry-After
header (which we expect to be the case for 429/Too Many Requests), the
delay it specifies will be respected. For other cases, it will back
off exponentially (after one immediate retry, doubling) up to 10 times,
with the library's maximum delay (120s).

Fixes #72