Support repeated keys in labelSelectors
To express the condition "foo!=bar,foo" (foo doesn't equal bar, but
does exist) isn't possible if we restrict the input to a dict.
This change means it can now either be a dict, or a list of
(key, value) tuples; in the case of a dict, it's converted to
such a list before it's processed.