discard watch event if Model instance can't be created
watch_list is a generator function intended to be used as a continuous
stream of updates on a resource kind.  When it fails to create an
instance of the Model type from the payload of a watch event because the
resource in the payload is not valid according to the Model class, a
TypeError is raised. This stops the stream of watch events and the
client has to try again/restart it by calling watch_list again. Since
the watch API returns all/most resources on the initial call, if the
resource that caused the exception to be raised is still present, the
same thing will happen again. In this case the client will not be able
to process all resources using watch_list until the invalid resource is
removed.

With this change, when creating a Model object from the payload of an
event raises TypeError, the exception is logged at ERROR level and the
event is discarded. This means that clients will be able to process all
valid resources if there is one or more resources that are not valid
according to the Model type. A drawback is that clients can't set their
own error handling for this failure mode. Since watch_list already does
the same thing if the payload of an event can not be parsed as JSON, and
a watch event without the object on it isn't particularly useful, I
think that is probably okay.

If it becomes necessary to support custom error handling later, one
option might perhaps be to let watch_list take function(s) as optional
parameters, which called in the except blocks for these failure
modes. The default can be the current behavior (logging the exception).