Treat read timeouts similar to the server hanging up the connection.
That is, stop yielding more data and return normally.
A warning is also logged however.

A read timeout might happen in empty clusters if there isn't any new
events appearing for the specified time. In most cases, the default
apiserver timeout will make it hang up the connection before that
happens, which is why this problem hasn't been noticed before.