Add resourceVersion handling to Watcher.
With this, I think a much lower stream timeout can be used, since new
calls should resume the watch in the same place. Also lowered the
logging message for timeout from warning to info due to this.

The exception is when there really aren't any new events for an extended
period, because then no new resourceVersion will be seen by the watcher.
Since the server will expire them fairly quickly each reconnect will do
a full quorum read when this is the case. Still, the cache should handle
that in most cases, so it seems acceptable.