Reduce risk of yielding stale events from watcher
It looks like bookmarks event aren't being sent on the watch connection
on the first connect, when the connection is started with
resourceVersion=None.

In some cases the lack of bookmark events can cause the watcher to yield
stale events after reconnecting for the first time. For example if we
assume a empty namespace where the following changes happen (in this
order):
- resource a is added with resourceVersion=1
- resource b is added with resourceVersion=2
- resource b is deleted

After that, the watch connection starts, and would yield one event for the only
existing resource;
- ADDED for resource a with resourceVersion=1

when the watch connection eventually times out and restarts it would use
resourceVersion=1, and yield
- ADDED   for resource b with resourceVersion=2
- DELETED for resource b with resourceVersion=2
since these events happened after resourceVersion 1.

To avoid these stale events; instead of relying on the initial list of
events from the watch_list call, do an actual list call before starting
watch connection, and use resourceVersion from the ListMeta returned
from list call when starting watch connection, as suggested in current
documentation[1].

This behavour could also be avoided by ensuring bookmark events are sent
on the initial watch connection, which could be achieved by enabling the
streaming lists[2] feature, but that is currenly only in alpha status as
of Kubernetes 1.27.

[1]: https://kubernetes.io/docs/reference/using-api/api-concepts/#efficient-detection-of-changes
[2]: https://kubernetes.io/docs/reference/using-api/api-concepts/#streaming-lists