Simplify WatchEvent and SyntheticAddedWatchEvent classes
- WatchEventBase.__init__ used to take event_json, but since it only needs
resourceVersion it has been changed to only take that as parameter,
and sublcasses are responsible for providing that value.

- WatchEvent.__init__ used to take event_json and Model class, this is now
supported via the WatchEvent.from_dict classmethod with the same
signature. WatchEvent.__init__ is changed to take WatchEvent type and
Model instance, to make it easier to extend, for example by
SyntheticAddedWatchEvent.